[RSDK-4484] Make an installer for board canaries (#2)
This is a shell script that installs the RDK server, a systemctl config and a crontab. The systemctl config will run the RDK server with the canary config (which you'll need to set up manually), and the crontab will run the hardware tests on this board every night at 2:30 AM.

Tried on an Orin, an Orin Nano, and an Up 4000: seems to work okay! 

Things that are still not included in this PR, and will be in a future one:
- Logging the results properly, and putting the logs somewhere convenient
- Monitoring the logs and sending notifications when something goes wrong


* create systemd config file

* add crontab template

* Add systemctl file for canaries

* first draft at installer script

* remove duplicate file: I forgot what I did yesterday

* Change crontab from 2:00 to 2:30: easier to remember which column is which later

* run the tests even if pulling the latest branch fails

* remove trailing whitespace

* log the date and store the output. We'll get proper logging soon

* add instructions for the manual steps, too

* Make install.sh executable

* Remember to install the RDK server

* debugging

* delimit regex with a non-slash so you can use slashes within

* tweak to output

* install the Python SDK, too

* don't fail if we try to stop a non-running server

* debugging