[RSDK-4766] give the option of specifying other pins for PWM/interrupts (#8)
Problem: the Beaglebone AI-64 pins seem to be either GPIO or PWM, but not both. The canary tests want to test both.

Solution: if you only define one set of pins in the test config, use those for everything. If you define a second set, though, do different tests on the different pins.

Tried on an Orin Nano with only 1 set of pins defined: the tests all pass.

Tried on a Beaglebone AI-64 with 2 sets of pins defined: the tests all pass again!


* give the option of specifying other pins for PWM/interrupts

* debugging

* more debugging

* rephrase comment

* Add template to example config