turn on Rust backtracing, in case we hit that panic again (#18)
This will help with https://viam.atlassian.net/browse/RSDK-6252 if it happens again.

Not tested: given that we can't reliably reproduce the issue, I don't even know how I'd go about testing it. but my expectation is that this is right, and then we'll get more details next time the panic happens.