Log to /var/log instead of /tmp (#19)
Because this changes the crontab template, we need to manually install this on all the boards. This kind of thing is what [Salt](https://docs.saltproject.io/en/master/topics/tutorials/walkthrough.html#salt-in-10-minutes/) is intended to do. but we have so few boards that it's not worth the hassle to set that up yet.

So, we shouldn't check this in until we're willing to manually update the crontabs on all the board canaries (likely by re-running `install.sh` on each of them). 

For the moment, I'm not adding in anything about log rotation. We add a couple kilobytes to the logs every night, which means we'll add less than a megabyte per year. If we rotate the logs once per decade, that's more than enough. :wink: 


* cron jobs should log to /var/log

* analyze logs from /var/log