Diagnose problems with counting ticks (#24)
Usually the canary tests pass fine, but sometimes the tick count for the HW PWM stuff is way too low. This PR prints out data about every interrupt we encounter, in the hopes that we can figure out what's going wrong.

Tried on an Orin Nano: works fine. We get lots of ticks printed out, and then the test passes. The cron job will pull the latest code from github, and then run `pip install -r requirements` to upgrade the SDK, so the boards should be able to upgrade themselves.


* print all data on all ticks

* debugging

* call stream_ticks

* bump requirements.txt to an SDK that supports streaming ticks

* whitespace tweak

* whitespace tweak