ignore the new extra output when checking if the tests have passed (#25)
This is a follow-up to #24, which added new output from the tests to try to better diagnose the flakiness around PWM output. What I forgot was that the later analysis checks that the test output looks right and sends it to Slack if it doesn't. When the test output changed, we started sending it all to Slack! This PR corrects it, so the analysis ignores the new output.

Tried on the Orin AGX: without this change it (erroneously) reports to Slack that last night's test failed, and with the change, it (correctly) says the tests passed.