logs analysis should include today's date, to make metaanalysis easier (#31)
This is split out of https://github.com/viamrobotics/board_canaries/pull/30, to get a head start. The idea is that if we log the current date when analyzing the logs, the monitor can check whether the canaries have analyzed their tests each day and write to Slack if something has gone wrong. but to prevent it from writing to Slack on the very first day, we should start logging the date before then.