make the crontabs use bash, not sh (#33)
Not described in this repo is the fact that the pi5 canary refuses to install any Python packages outside of a virtual environment. So, on the pi5 I installed everything inside one, and then edited the repo on that board to source the venv before running the Python scripts. Specifically, I added the line `source venv/bin/activate` inside both `analyze_logs.sh` and `canary_tests.sh`. That all worked fine. 

When the pi5 became the canary monitor, I also added that line inside the new cron entry, but didn't notice that cron was using `sh` instead of `bash` as its shell! and `sh` does not define `source`, so sourcing the venv was failing as an unknown command. The two files mentioned previously both start with `#!/bin/bash`, which is why they worked but the cron job didn't!

Thanks to Dan G for the help figuring this out.

Tried on the pi5 canary monitor: the cron job ran and passed last night!